BUG#30538516 METHODS WHICH DEAL WITH FILESYSTEM DO NOT PROPERLY HANDLE NON-ASCII CHARACTERS
MySQL Shell strings are UTF-8 encoded. NTFS, Windows file system, stores
filenames in UTF-16 encoding. All operations on NTFS file system must be
translated from UTF-8 to UTF-16, and all strings obtained from Unicode
function versions of Windows API calls must be converted back to UTF-8.

+++ mysqlshdk/libs/db/mysqlx/tokenizer.h
@@ -152,7 +152,7 @@ class Tokenizer {
   const Token &peek_token();
   void unget_token();
   void inc_pos_token();
-  int get_token_pos() { return _pos; }
+  int get_token_pos() { return static_cast<int>(_pos); }
   const Token &consume_any_token();
   void assert_tok_position();
   bool tokens_available();
